requirements.txt
anthropic>=0.18.0
click>=8.1.0
pyyaml>=6.0
colorama>=0.4.6
python-dotenv>=1.0.0
